✨ Implement eigendecomposition utilities for real and complex matrices; add corresponding tests for new functionality
