+++ mlir/lib/Dialect/QCO/Utils/CMakeLists.txt
@@ -6,7 +6,14 @@
 #
 # Licensed under the MIT License
 
-add_mlir_library(MLIRQCOMatrix PARTIAL_SOURCES_INTENDED Matrix.cpp LINK_LIBS PUBLIC MLIRSupport)
+add_mlir_library(
+  MLIRQCOMatrix
+  PARTIAL_SOURCES_INTENDED
+  Matrix.cpp
+  Eigensolver.cpp
+  LINK_LIBS
+  PUBLIC
+  MLIRSupport)
 
 mqt_mlir_target_use_project_options(MLIRQCOMatrix)
 
@@ -14,7 +21,7 @@ target_sources(MLIRQCOMatrix PUBLIC FILE_SET HEADERS BASE_DIRS ${MQT_MLIR_SOURCE
                                     ${MQT_MLIR_SOURCE_INCLUDE_DIR}/mlir/Dialect/QCO/Utils/Matrix.h)
 
 file(GLOB_RECURSE UTILS_CPP "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
-list(FILTER UTILS_CPP EXCLUDE REGEX "Matrix\\.cpp$")
+list(FILTER UTILS_CPP EXCLUDE REGEX "(Matrix|Eigensolver)\\.cpp$")
 
 add_mlir_dialect_library(
   MLIRQCOUtils
